correções após reiview

+++ fronend/src/pages/CadastroTurma/CadastroTurma.jsx
@@ -9,7 +9,7 @@ function CadastroTurma() {
 
   const [turmaData, setTurmaData] = useState({});
   const [turma, setTurma] = useState([])
-
+  const [turmaNome, setTurmaNome] =  useState('');
 
   const handleSubmit = async (event)=>{
     event.preventDefault();
@@ -60,8 +60,18 @@ useEffect(()=>{
     
 }
 const handleDelete = async (codigo) =>{
-  await turmaService.deleteTurma(codigo);
-  await carregaTurma();
+  const confirmacao = window.confirm("Confirma a exclusão?");
+
+  if(confirmacao){
+    try {
+       await turmaService.deleteTurma(codigo);
+       await carregaTurma();
+       alert("Turma excluída com sucesso!")
+    } catch (error) {
+        alert("Erro ao excluir turma!!")
+    }
+  }
+     
 }
 
 
@@ -89,6 +99,44 @@ const atualizarTurma = async () => {
     alert('Erro ao atualizar! ')
     console.log('Erro ao atualizar: ', error)
   }
+}
+
+
+async function getByNome(nomee) {
+
+  const nome = {
+    nome: `${nomee}`
+  }
+
+
+  try {
+
+    const dados = await turmaService.filtrar(nome)
+
+    if (dados.length > 0) {
+
+      const dadosFiltroNome = dados.map((turma) => (
+        {
+          codigo: `${turma.codigo}`,
+          descricao: `${turma.descricao}`,
+          anoTurma: `${turma.anoTurma}`,
+          qtde: `${turma.qtde}`,
+          
+        }
+      ));
+       
+      setTurma(dadosFiltroNome);
+  
+    }
+    else {
+      
+    }
+    
+  }
+  catch (erro) {
+
+  }
+  
 }
 
 
@@ -191,6 +239,23 @@ const atualizarTurma = async () => {
       </div>
     </form>
           <div class="container mt-4">
+          <div class="row">
+              <div class="input-group flex-nowrap">
+                <input
+                  name="nome"
+                  id="nome"
+                  type="text"
+                  class="form-control"
+                  placeholder="Pesquisar por nome"
+                  value={turma.nome}
+                  onChange={handleInputChange}
+                  onBlur={(e) => setTurmaNome(e.target.value)}
+                  required
+                />
+                  &nbsp; &nbsp;
+                <i class="bi bi-search my-custom-icon"  onClick={()=>getByNome(turmaNome)} ></i>
+            </div>
+            </div>
           <div class="table-responsive">
               <table class="table">
                   <thead>
